Split hosted cutover runtime boundaries
